chore: address #279 review — extract test helpers and unify unique_name
Adds unique_name, unique_did, unique_subject, and test_credential to
dialog-storage's shared helpers module (gated behind the existing
'helpers' feature). Replaces the per-file duplicates in fs/, indexeddb/,
and volatile/ test modules.

Effects on the unique_name pattern:
- IndexedDb tests previously used a bare js_sys::Date::now() (millisecond
  resolution). They now use the timestamp + atomic counter pattern that
  the FS tests already used, eliminating same-millisecond collisions when
  multiple tests run concurrently.
- Native side keeps high-resolution dialog_common::time. Wasm side scales
  Date::now() ms to ns so the value matches the native scale.

Other small cleanups:
- Drop now-unused js_sys::Date import from indexeddb.rs tests.
- Replace the inline Date::now()-based ts in it_isolates_data_across_locations
  with unique_name to avoid the same collision risk.
