+++ rust/dialog-storage/src/storage/provider/volatile/archive.rs
@@ -65,22 +65,8 @@ impl Provider<Put> for Volatile {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use dialog_capability::{Did, Subject};
+    use crate::helpers::unique_subject;
     use dialog_effects::archive::{Archive, Catalog};
-
-    fn unique_subject(prefix: &str) -> Subject {
-        let did: Did = format!(
-            "did:test:{}-{}",
-            prefix,
-            dialog_common::time::now()
-                .duration_since(dialog_common::time::UNIX_EPOCH)
-                .unwrap()
-                .as_nanos()
-        )
-        .parse()
-        .unwrap();
-        Subject::from(did)
-    }
 
     #[dialog_common::test]
     async fn it_returns_none_for_missing_content() -> anyhow::Result<()> {
